requirements.txt
asttokens==2.4.1
attrs==23.2.0
blobfile==2.1.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
diskcache==5.6.3
executing==2.0.1
fairscale==0.4.13
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.24.0
idna==3.7
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
json_strong_typing==0.3.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
llama_cpp_python==0.2.83
llama_models==0.0.1
lxml==4.9.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.0.1
packaging==24.1
parso==0.8.4
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pycryptodomex==3.20.0
pydantic==1.10.13
pydantic_core==2.18.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rpds-py==0.19.0
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tbb==2021.13.0
tiktoken==0.7.0
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
